pyproject.toml
[project]
name = "expert-rating"
version = "0.1.0"
description = ""
authors = [
    {name = "niktaparkur",email = "153425457+niktaparkur@users.noreply.github.com"}
]
packages = [{include = "src"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11, <4"
dependencies = [
    "fastapi[all] (>=0.118.0,<0.119.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "aionvk (>=0.4.13,<0.5.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.1,<47.0.0)",
    "booktype (>=1.5,<2.0)",
    "redis[hiredis] (==5.0.1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.13.2,<0.14.0)"
]
